added SERIAL inner-field OP support BY_UNIQUE&COND
includes:
 added DB::Cell::Serial::Value::FieldUpdate::CTRL_VALUE_{SET,DEL}

[TEST COMMIT]

+++ src/cc/lib/swcdb/db/Cells/CellValueSerialField.cc
@@ -198,7 +198,7 @@ void Field_LIST_INT64::print(std::ostream& out) const {
   if(size) {
     const uint8_t* ptr = base;
     for(size_t remain = size;;) {
-      out << Serialization::decode_vi64(&ptr, &remain);
+      out << int64_t(Serialization::decode_vi64(&ptr, &remain));
       if(!remain)
         break;
       out << ',';
